Refactored how filtering works. Filtering exclusively happens on Adze logs rather than Shed. To apply filters from a Shed, you just add the filters to the global override config.
