+++ src/_defaults/shed_defaults.ts
@@ -3,5 +3,4 @@ import { ShedConfig } from '../_contracts';
 export const shed_defaults: ShedConfig = {
   cache_limit: 300,
   global_cfg: null,
-  filters: {},
 };
